adds second unit test file for test runner
